dashboard1, radio, graphics update

+++ src/com/ezio/multiwii/mw/MultirotorData.java
@@ -166,7 +166,7 @@ public abstract class MultirotorData {
 
 	public float mot[] = new float[8];
 	public float servo[] = new float[8];
-	public float rcThrottle = 1500, rcRoll = 1500, rcPitch = 1500, rcYaw = 1500, rcAUX1 = 1500, rcAUX2 = 1500, rcAUX3 = 1500, rcAUX4 = 1500;
+	public int rcThrottle = 1500, rcRoll = 1500, rcPitch = 1500, rcYaw = 1500, rcAUX1 = 1500, rcAUX2 = 1500, rcAUX3 = 1500, rcAUX4 = 1500;
 	public int SensorPresent = 0;
 	public int mode = 0;
